fixed keyboard movement (float)

+++ player/Player.java
@@ -16,8 +16,8 @@ public abstract class Player {
     protected boolean isActive;
     protected boolean hasImmunity;
     protected long lastImmunity;
-    protected int x;
-    protected int y;
+    protected float x;
+    protected float y;
     protected EnemyDeletable parent;
     private int senbonSakuraN;
     private int senbonSakuraC;
@@ -35,11 +35,11 @@ public abstract class Player {
     }
 
     public int getX() {
-        return x;
+        return (int)x;
     }
 
     public int getY() {
-        return y;
+        return (int)y;
     }
 
     public int getLives() {
@@ -65,7 +65,7 @@ public abstract class Player {
         lives = _lives;
     }
 
-    public void decLives(int _x, int _y) {
+    public void decLives(float _x, float _y) {
         lives--;
         x = _x;
         y = _y;
@@ -97,10 +97,10 @@ public abstract class Player {
 
     public void paint(Graphics g) {
         g.setColor(Color.WHITE);
-        g.fillOval(x - 5, y - 5, 10, 10);
+        g.fillOval((int)x - 5, (int)y - 5, 10, 10);
         if (senbonSakuraC-->0) {
             g.setColor(Color.PINK);
-            g.fillOval(x-SENBONSAKURA_RADIUS/2, y-SENBONSAKURA_RADIUS/2, SENBONSAKURA_RADIUS, SENBONSAKURA_RADIUS);
+            g.fillOval((int)x-SENBONSAKURA_RADIUS/2, (int)y-SENBONSAKURA_RADIUS/2, SENBONSAKURA_RADIUS, SENBONSAKURA_RADIUS);
         }
         int barWidth = (int)Math.min(SENBONSAKURA_TIMEOUT, System.currentTimeMillis()-senbonSakuraT);
         barWidth *=.05;
